// packages/admin-sdk/src/zod/generated/Collection.ts
// This file is auto-generated. Do not edit directly.
import { z } from 'zod';
import { CollectionRuleSchema } from './CollectionRule';
import { CustomFieldSchema } from './CustomField';

export const CollectionSchema = z.object({
  id: z.string(),
  name: z.string(),
  permalink: z.string(),
  position: z.number(),
  sort_order: z.string(),
  meta_title: z.string().nullable(),
  meta_description: z.string().nullable(),
  meta_keywords: z.string().nullable(),
  products_count: z.number(),
  description: z.string(),
  description_html: z.string(),
  image_url: z.string().nullable(),
  square_image_url: z.string().nullable(),
  custom_fields: z.array(CustomFieldSchema).optional(),
  translations: z.record(z.string(), z.record(z.string(), z.union([z.string(), z.number()]).nullable())).optional(),
  automatic: z.boolean(),
  rules_match_policy: z.string(),
  metadata: z.record(z.string(), z.unknown()),
  created_at: z.string(),
  updated_at: z.string(),
  rules: z.array(CollectionRuleSchema).optional(),
});

export type Collection = z.infer<typeof CollectionSchema>;